Move templates for ptq tests to separate directory
