+++ tests/onnx/quantization/test_calculation_quantizer_params.py
@@ -15,7 +15,7 @@ import pytest
 from nncf.onnx.quantization.quantizer_parameters import calculate_scale_zero_point
 from nncf.onnx.quantization.quantizer_parameters import get_level_low_level_high
 from nncf.onnx.statistics.collectors import ONNXMinMaxTensorStatistic
-from tests.post_training.test_calculate_quantizer_parameters import TemplateTestFQParams
+from tests.post_training.test_templates.test_calculate_quantizer_parameters import TemplateTestFQParams
 
 
 @pytest.mark.parametrize(
