+++ tests/openvino/native/quantization/test_ptq_params.py
@@ -34,8 +34,8 @@ from tests.common.quantization.metatypes import LinearTestMetatype
 from tests.common.quantization.metatypes import SoftmaxTestMetatype
 from tests.openvino.native.models import DepthwiseConv4DModel
 from tests.openvino.native.models import LinearModel
-from tests.post_training.models import NNCFGraphToTestMatMul
-from tests.post_training.test_ptq_params import TemplateTestPTQParams
+from tests.post_training.test_templates.models import NNCFGraphToTestMatMul
+from tests.post_training.test_templates.test_ptq_params import TemplateTestPTQParams
 
 
 def get_patterns_setup() -> GraphPattern:
