+++ tests/torch/ptq/test_calculation_quantizer_params.py
@@ -33,7 +33,7 @@ from nncf.quantization.fake_quantize import get_quantizer_narrow_range
 from nncf.torch.model_creation import create_nncf_network
 from nncf.torch.statistics.aggregator import PTStatisticsAggregator
 from nncf.torch.tensor_statistics.statistics import PTMinMaxTensorStatistic
-from tests.post_training.test_calculate_quantizer_parameters import TemplateTestFQParams
+from tests.post_training.test_templates.test_calculate_quantizer_parameters import TemplateTestFQParams
 from tests.torch.helpers import get_all_inputs_for_graph_node
 from tests.torch.helpers import get_nodes_by_type
